user role

+++ resources/views/users/index.blade.php
@@ -29,7 +29,9 @@
     <div class="panel-heading">
       <h5 class="panel-title"></h5>
       <div class="heading-elements"> </div>
+      @can('user-create')
       <div class="pull-right"> <a class="btn btn-success" href="{{ route('users.create') }}"> Create New User</a> </div>
+      @endcan
     </div>
     <div class="panel-body">
       <div class="row"> 
@@ -63,10 +65,14 @@
     </td>
     <td>
        <a class="btn btn-info" href="{{ route('users.show',$user->id) }}">Show</a>
+       @can('user-edit')
        <a class="btn btn-primary" href="{{ route('users.edit',$user->id) }}">Edit</a>
+       @endcan
+       @can('user-delete')
         {!! Form::open(['method' => 'DELETE','route' => ['users.destroy', $user->id],'style'=>'display:inline']) !!}
             {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
         {!! Form::close() !!}
+        @endcan
     </td>
   </tr>
  @endforeach
